// src/typings/agora-rtm-sdk.d.ts
declare namespace RtmStatusCode {
  enum ConnectionChangeReason {
    LOGIN = "LOGIN",
    LOGIN_SUCCESS = "LOGIN_SUCCESS",
    LOGIN_FAILURE = "LOGIN_FAILURE",
    LOGIN_TIMEOUT = "LOGIN_TIMEOUT",
    INTERRUPTED = "INTERRUPTED",
    LOGOUT = "LOGOUT",
    BANNED_BY_SERVER = "BANNED_BY_SERVER",
    REMOTE_LOGIN = "REMOTE_LOGIN",
  }
  enum ConnectionState {
    DISCONNECTED = "DISCONNECTED",
    CONNECTING = "CONNECTING",
    CONNECTED = "CONNECTED",
    RECONNECTING = "RECONNECTING",
    ABORTED = "ABORTED",
  }
  enum LocalInvitationState {
    IDLE = "IDLE",
    SENT_TO_REMOTE = "SENT_TO_REMOTE",
    RECEIVED_BY_REMOTE = "RECEIVED_BY_REMOTE",
    ACCEPTED_BY_REMOTE = "ACCEPTED_BY_REMOTE",
    REFUSED_BY_REMOTE = "REFUSED_BY_REMOTE",
    CANCELED = "CANCELED",
    FAILURE = "FAILURE",
  }
  enum RemoteInvitationState {
    INVITATION_RECEIVED = "INVITATION_RECEIVED",
    ACCEPT_SENT_TO_LOCAL = "ACCEPT_SENT_TO_LOCAL",
    REFUSED = "REFUSED",
    ACCEPTED = "ACCEPTED",
    CANCELED = "CANCELED",
    FAILURE = "FAILURE",
  }
  enum LocalInvitationFailureReason {
    UNKNOWN = "UNKNOWN",
    PEER_NO_RESPONSE = "PEER_NO_RESPONSE",
    INVITATION_EXPIRE = "INVITATION_EXPIRE",
    PEER_OFFLINE = "PEER_OFFLINE",
    NOT_LOGGEDIN = "NOT_LOGGEDIN",
  }
  enum RemoteInvitationFailureReason {
    UNKNOWN = "UNKNOWN",
    PEER_OFFLINE = "PEER_OFFLINE",
    ACCEPT_FAILURE = "ACCEPT_FAILURE",
    INVITATION_EXPIRE = "INVITATION_EXPIRE",
  }
  enum PeerOnlineState {
    ONLINE = "ONLINE",
    UNREACHABLE = "UNREACHABLE",
    OFFLINE = "OFFLINE",
  }
  enum PeerSubscriptionOption {
    ONLINE_STATUS = "ONLINE_STATUS",
  }
  enum MessageType {
    TEXT = "TEXT",
    RAW = "RAW",
    IMAGE = "IMAGE",
    FILE = "FILE",
  }
  enum LegacyAreaCode {
    CN = "CN",
    NA = "NA",
    EU = "EU",
    AS = "AS",
    JP = "JP",
    IN = "IN",
    GLOB = "GLOB",
    OC = "OC",
    SA = "SA",
    AF = "AF",
    OVS = "OVS",
  }
  enum AreaCode {
    GLOBAL = "GLOBAL",
    INDIA = "INDIA",
    JAPAN = "JAPAN",
    ASIA = "ASIA",
    EUROPE = "EUROPE",
    CHINA = "CHINA",
    NORTH_AMERICA = "NORTH_AMERICA",
  }
}

interface ChannelAttributeProperties {
  value: string;

  lastUpdateUserId: string;

  lastUpdateTs: number;
}

interface AttributesMap {
  [key: string]: string;
}
interface ChannelAttributes {
  [key: string]: ChannelAttributeProperties;
}

interface ChannelAttributeOptions {
  enableNotificationToChannelMembers?: boolean;
}

declare type ListenerType<T> = [T] extends [(...args: infer U) => any]
  ? U
  : [T] extends [void]
  ? []
  : [T];
declare class EventEmitter<TEventRecord = {}> {
  static defaultMaxListeners: number;
  on<P extends keyof TEventRecord, T>(
    this: T,
    event: P,
    listener: (...args: ListenerType<TEventRecord[P]>) => void
  ): this;

  once<P extends keyof TEventRecord, T>(
    this: T,
    event: P,
    listener: (...args: ListenerType<TEventRecord[P]>) => void
  ): this;

  off<P extends keyof TEventRecord, T>(
    this: T,
    event: P,
    listener: (...args: any[]) => any
  ): this;

  removeAllListeners<P extends keyof TEventRecord, T>(this: T, event?: P): this;
  listeners<P extends keyof TEventRecord, T>(this: T, event: P): Function[];
  rawListeners<P extends keyof TEventRecord, T>(this: T, event: P): Function[];
  listenerCount<P extends keyof TEventRecord, T>(this: T, event: P): number;
}

interface RtmTextMessage {
  text: string;

  messageType?: "TEXT";
  rawMessage?: never;
  description?: never;
}

interface RtmRawMessage {
  rawMessage: Uint8Array;

  description?: string;

  messageType?: "RAW";
  text?: never;
}

interface RtmImageMessage {
  width: number;

  height: number;

  fileName: string;

  description: string;

  thumbnail: Blob | undefined;

  thumbnailWidth: number;

  thumbnailHeight: number;

  size: number;

  mediaId: string;

  messageType: "IMAGE";
}

interface RtmFileMessage {
  fileName: string;

  description: string;

  thumbnail: Blob | undefined;

  size: number;

  mediaId: string;

  messageType: "FILE";
}

type RtmMessage =
  | RtmTextMessage
  | RtmRawMessage
  | RtmFileMessage
  | RtmImageMessage;

interface PeerMessageSendResult {
  hasPeerReceived: boolean;
}

interface SendMessageOptions {
  enableOfflineMessaging?: boolean;
  enableHistoricalMessaging?: boolean;
}

interface ReceivedMessageProperties {
  serverReceivedTs: number;
  isOfflineMessage: boolean;
}

interface PeersOnlineStatusMap {
  [peerId: string]: RtmStatusCode.PeerOnlineState;
}

declare namespace RtmEvents {
  export interface RtmChannelEvents {
    ChannelMessage: (
      message: RtmMessage,
      memberId: string,
      messagePros: ReceivedMessageProperties
    ) => void;

    MemberLeft: (memberId: string) => void;

    MemberJoined: (memberId: string) => void;

    AttributesUpdated: (attributes: ChannelAttributes) => void;
    MemberCountUpdated: (memberCount: number) => void;
  }

  export interface RemoteInvitationEvents {
    RemoteInvitationCanceled: (content: string) => void;

    RemoteInvitationRefused: () => void;

    RemoteInvitationAccepted: () => void;

    RemoteInvitationFailure: (
      reason: RtmStatusCode.RemoteInvitationFailureReason
    ) => void;
  }

  export interface LocalInvitationEvents {
    LocalInvitationAccepted: (response: string) => void;
    LocalInvitationRefused: (response: string) => void;
    LocalInvitationReceivedByPeer: () => void;
    LocalInvitationCanceled: () => void;
    LocalInvitationFailure: (
      reason: RtmStatusCode.LocalInvitationFailureReason
    ) => void;
  }

  export interface RtmClientEvents {
    MessageFromPeer: (
      message: RtmMessage,
      peerId: string,
      messageProps: ReceivedMessageProperties
    ) => void;
    ConnectionStateChanged: (
      newState: RtmStatusCode.ConnectionState,
      reason: RtmStatusCode.ConnectionChangeReason
    ) => void;
    RemoteInvitationReceived: (remoteInvitation: RemoteInvitation) => void;

    TokenExpired: () => void;

    PeersOnlineStatusChanged: (status: PeersOnlineStatusMap) => void;

    FallbackProxyConnected: () => void;
  }
}
// ...
declare class RemoteInvitation extends EventEmitter<RtmEvents.RemoteInvitationEvents> {
  readonly channelId: string;

  readonly callerId: string;

  readonly content: string;

  readonly state: RtmStatusCode.RemoteInvitationState;

  response: string;

  accept(): void;

  refuse(): void;

  on<EventName extends keyof RtmEvents.RemoteInvitationEvents>(
    eventName: EventName,
    listener: (
      ...args: ListenerType<RtmEvents.RemoteInvitationEvents[EventName]>
    ) => any
  ): this;
}